misc(feat): moonshot-v1-8k
